Implemented path cost function and modified related methods to adapt with the new return type.

+++ src/code/Problem/Problem.java
@@ -10,7 +10,7 @@ public abstract class Problem {
         this.operators = operators;
     }
 
-    public abstract int pathCost(State state);
+    public abstract double pathCost(State state);
 
     public abstract boolean isGoalState(State state);
